Small Refactoring
