+++ src/components/game-screen/game-screen.jsx
@@ -4,9 +4,9 @@ import {connect} from "react-redux";
 import {GameType, MAX_MISTAKE_COUNT} from "../../utils/const";
 import MusicianQuestionScreen from "../musician-question-screen/musician-question-screen";
 import GenreQuestionSreen from "../genre-question-screen/genre-question-screen";
-import {gameScreenType} from "../types/types";
-import withAudioPlayer from "../hocs/with-audio-player/with-audio-player";
-import withUserAnswer from "../hocs/with-user-answer/with-user-answer";
+import {gameScreenType} from "../../types/types";
+import withAudioPlayer from "../../hocs/with-audio-player/with-audio-player";
+import withUserAnswer from "../../hocs/with-user-answer/with-user-answer";
 import {ActionCreator} from "../../store/action";
 import Mistakes from "../mistakes/mistakes";
